Move files related to ADQ patch to appropriate directory
