+++ src/solver/cmake/solver.cmake
@@ -54,16 +54,16 @@ set(RTESOLVER_OPT
 		optimisation/opt_export_structure.cpp
 		optimisation/base_weekly_optimization.h
 		optimisation/base_weekly_optimization.cpp
-		optimisation/adequacy_patch_weekly_optimization.h
-		optimisation/adequacy_patch_weekly_optimization.cpp
+		optimisation/adequacy_patch_local_matching/adequacy_patch_weekly_optimization.h
+		optimisation/adequacy_patch_local_matching/adequacy_patch_weekly_optimization.cpp
 		optimisation/weekly_optimization.h
 		optimisation/weekly_optimization.cpp
 		optimisation/optim_post_process_list.h
 		optimisation/optim_post_process_list.cpp
-		optimisation/adq_patch_post_process_list.h
-		optimisation/adq_patch_post_process_list.cpp
 		optimisation/post_process_commands.h
 		optimisation/post_process_commands.cpp
+		optimisation/adequacy_patch_csr/adq_patch_post_process_list.h
+		optimisation/adequacy_patch_csr/adq_patch_post_process_list.cpp
 		optimisation/adequacy_patch_local_matching/adq_patch_local_matching.h
 		optimisation/adequacy_patch_local_matching/adq_patch_local_matching.cpp
 		optimisation/adequacy_patch_csr/adq_patch_curtailment_sharing.h
