+++ src/solver/optimisation/adequacy_patch_csr/adq_patch_post_process_list.cpp
@@ -1,5 +1,5 @@
 #include "adq_patch_post_process_list.h"
-#include "post_process_commands.h"
+#include "../post_process_commands.h"
 
 
 namespace Antares::Solver::Simulation
