+++ src/solver/optimisation/adequacy_patch_csr/adq_patch_post_process_list.h
