+++ src/solver/optimisation/adequacy_patch_csr/construct_problem_variables.cpp
@@ -28,7 +28,7 @@
 #include "../solver/optimisation/opt_structure_probleme_a_resoudre.h"
 #include "../solver/simulation/simulation.h"
 #include "../solver/simulation/sim_structure_donnees.h"
-#include "../hourly_csr_problem.h"
+#include "hourly_csr_problem.h"
 
 #include "pi_constantes_externes.h"
 void HourlyCSRProblem::constructVariableENS()
