+++ src/solver/optimisation/adequacy_patch_csr/hourly_csr_problem.h
