+++ src/solver/optimisation/adequacy_patch_local_matching/adequacy_patch_weekly_optimization.h
@@ -27,7 +27,7 @@
 
 #pragma once
 
-#include "base_weekly_optimization.h"
+#include "../base_weekly_optimization.h"
 #include "../simulation/sim_structure_probleme_economique.h"
 #include "antares/study/area/area.h"
