+++ src/solver/optimisation/opt_fonctions.h
@@ -30,7 +30,7 @@
 #include "../config.h"
 #include "opt_structure_probleme_a_resoudre.h"
 #include "../simulation/sim_structure_donnees.h"
-#include "hourly_csr_problem.h"
+#include "adequacy_patch_csr/hourly_csr_problem.h"
 #include "opt_period_string_generator_base.h"
 
 void OPT_OptimisationHebdomadaire(PROBLEME_HEBDO*, uint);
