+++ src/solver/optimisation/post_process_commands.cpp
@@ -1,7 +1,6 @@
 
 #include "post_process_commands.h"
 #include "../simulation/common-eco-adq.h"
-#include "adequacy_patch_weekly_optimization.h"
 #include "adequacy_patch_csr/adq_patch_curtailment_sharing.h"
 
 namespace Antares::Solver::Simulation
