+++ src/solver/simulation/base_post_process.cpp
@@ -2,7 +2,7 @@
 #include <memory>
 #include "base_post_process.h"
 #include "../optimisation/optim_post_process_list.h"
-#include "../optimisation/adq_patch_post_process_list.h"
+#include "../optimisation/adequacy_patch_csr/adq_patch_post_process_list.h"
 
 namespace Antares::Solver::Simulation
 {
